requirements.txt
PySide2~=5.15
pdpyras
pyyaml
